Refactor optimize indexes
