+++ src/containers/Optimize/hooks/useGetOptimizedLineupsMutationResponse.ts
@@ -1,6 +1,6 @@
 import { useGetOptimizedLineupsMutation } from 'src/api';
 
-export default function useGetOptimizedLineupsMutationResponse() {
+export function useGetOptimizedLineupsMutationResponse() {
 	return useGetOptimizedLineupsMutation({
 		fixedCacheKey: 'optimize',
 	});
