+++ src/containers/Optimize/index.ts
@@ -1,34 +1,8 @@
-import { getOptimizedLineups } from './api/getOptimizedLineups';
-import Optimize from './components/Optimize.component';
-import optimizeHandler from './handlers/optimize.handler';
-import useGetOptimizedLineupsMutationResponse from './hooks/useGetOptimizedLineupsMutationResponse';
-import IGetOptimizedLineupsResponse from './models/IGetOptimizedLineupsResponse';
-import IOptimizedLineup from './models/IOptimizedLineup';
-import OptimizeReducers, {
-	optimizedState,
-	setLockedPlayers,
-} from './redux/Optimize.reducers';
-
-// Api
-export { getOptimizedLineups };
-
-// Handlers
-export { optimizeHandler };
-
-// Reducers
-export { OptimizeReducers };
-
-// State
-export { optimizedState };
-
-// Models
-export type { IGetOptimizedLineupsResponse, IOptimizedLineup };
-
-// Hooks
-export { useGetOptimizedLineupsMutationResponse };
-
-// Actions
-export { setLockedPlayers };
-
-// Components
-export default Optimize;
+export * from './api';
+export * from './components';
+export * from './handlers';
+export * from './hooks';
+export * from './mocks';
+export * from './models';
+export * from './reducers';
+export * from './settings';
