+++ src/containers/Optimize/models/IGetOptimizedLineupsPostBody.ts
@@ -2,7 +2,7 @@ import { IPlayer } from '@/containers/Players';
 import { EProviders } from '@/containers/Providers';
 import { ISport } from '@/containers/Sports';
 
-import IOptimizeSettings from './IOptimizeSettings';
+import { IOptimizeSettings } from './IOptimizeSettings';
 
 export interface IGetOptimizedLineupsPostBody {
 	players: IPlayer[];
