+++ src/containers/Optimize/models/index.ts
@@ -1,0 +1,4 @@
+export * from './IGetOptimizedLineupsPostBody';
+export * from './IGetOptimizedLineupsResponse';
+export * from './IOptimizedLineup';
+export * from './IOptimizeSettings';
