+++ src/containers/Optimize/settings/Optimize.settings.tsx
@@ -1,11 +1,9 @@
 import OptimizeNumberOfGenerations from './components/Optimize.numberOfGenerations';
 import OptimizeStatusFilters from './components/Optimize.statusFilters';
 
-const OptimizeSettings = () => (
+export const OptimizeSettings = () => (
 	<div className="space-y-6">
 		<OptimizeNumberOfGenerations />
 		<OptimizeStatusFilters />
 	</div>
 );
-
-export default OptimizeSettings;
